+++ www/Library/layouts/Admin/ProgramsMenu/ProgramsMenu.php
@@ -2,47 +2,44 @@
 /**
  * Меню программ
  * Меню автоматически формируется в зависимости от отображаемого объекта и доступного для него программ
- * @version 1.0
+ * @version 2.0
  */
 namespace Library\layouts\Admin\ProgramsMenu;
 
-use Library\views\Widget\Widget,
-    Boolive\values\Rule;
+use Library\views\Widget\Widget;
 
 class ProgramsMenu extends Widget
 {
+//    public function defineInputRule()
+//    {
+//        $this->_input_rule = Rule::arrays(array(
+//                'REQUEST' => Rule::arrays(array(
+//                        'object' => Rule::any(
+//                            Rule::arrays(Rule::entity()),
+//                            Rule::entity()
+//                        )->required()
+//                    )
+//                )
+//            )
+//        );
+//    }
 
-    public function defineInputRule()
-    {
-        $this->_input_rule = Rule::arrays(array(
-                'REQUEST' => Rule::arrays(array(
-                        'object' => Rule::any(
-                            Rule::arrays(Rule::entity()),
-                            Rule::entity()
-                        )->required()
-                    )
-                )
-            )
-        );
+    protected function initInputChild($input){
+        parent::initInputChild($input);
+        // Отображется список видов
+        $this->_input_child['REQUEST']['program'] = $this->programs->linked()->views;
+        // Начальная чаcть uri, которая будет отрезаться для получения относительного пути на вид
+        $this->_input_child['REQUEST']['base_uri'] = $this->_input_child['REQUEST']['program']->uri().'/';
+        // Не показыват корневой пункт меню
+        $this->_input_child['REQUEST']['show'] = false;
+        // Не использовать явное указание, каким видом показывать пункт меню
+        if (isset($this->_input_child['REQUEST']['view_name'])){
+            unset($this->_input_child['REQUEST']['view_name']);
+        }
     }
 
-    public function work($v = array())
-    {
-        $case = $this->programs->linked()->switch_views->getCase($this->_commands, $this->_input);
-        $programs = $case->find();
-        $c = new \Boolive\commands\Commands();
-        $v['items'] = array();
-        foreach ($programs as $p){
-            $p = $p->linked();
-            if ($p instanceof Widget && $p->canWork($c, $this->_input_child)){
-                $item = array(
-                    'title' => $p->title->value(),
-                    'view_name' => $p->name(),
-                    'icon' => $p->icon->file()
-                );
-                $v['items'][] = $item;
-            }
-        }
+    public function work($v = array()){
+        $v['item_view'] = $this->startChild('item_view');
         return parent::work($v);
     }
 }